+++ test/unit/org/apache/shiro/cas/grails/ShiroCasConfigUtilsSpec.groovy
@@ -2,7 +2,11 @@ package org.apache.shiro.cas.grails
 
 import grails.util.Holders
 import org.apache.commons.logging.Log
+import org.apache.shiro.SecurityUtils
+import org.apache.shiro.web.util.WebUtils
 import spock.lang.Specification
+
+import javax.servlet.http.HttpServletRequest
 
 class ShiroCasConfigUtilsSpec extends Specification {
     Log realLog
@@ -28,8 +32,8 @@ class ShiroCasConfigUtilsSpec extends Specification {
         and: "a default (but non-working) configuration is used"
         ShiroCasConfigUtils.serverUrl == ""
         ShiroCasConfigUtils.serviceUrl == ""
-        ShiroCasConfigUtils.loginUrl == ""
-        ShiroCasConfigUtils.logoutUrl == ""
+        ShiroCasConfigUtils.defaultLoginUrl == ""
+        ShiroCasConfigUtils.defaultLogoutUrl == ""
         ShiroCasConfigUtils.failureUrl == null
         ShiroCasConfigUtils.shiroCasFilter == "/shiro-cas=singleSignOutFilter,casFilter\n"
         !ShiroCasConfigUtils.singleSignOutDisabled
@@ -52,8 +56,8 @@ security.shiro.cas.serviceUrl = "http://localhost:8080/app/shiro-cas"
         ShiroCasConfigUtils.serviceUrl == "http://localhost:8080/app/shiro-cas"
 
         and: "other values are either defaulted or based on the configured values"
-        ShiroCasConfigUtils.loginUrl == "https://localhost/cas/login?service=http://localhost:8080/app/shiro-cas"
-        ShiroCasConfigUtils.logoutUrl == "https://localhost/cas/logout?service=http://localhost:8080/app/shiro-cas"
+        ShiroCasConfigUtils.defaultLoginUrl == "https://localhost/cas/login?service=http://localhost:8080/app/shiro-cas"
+        ShiroCasConfigUtils.defaultLogoutUrl == "https://localhost/cas/logout?service=http://localhost:8080/app/shiro-cas"
         ShiroCasConfigUtils.failureUrl == null
         ShiroCasConfigUtils.shiroCasFilter == "/shiro-cas=singleSignOutFilter,casFilter\n"
         !ShiroCasConfigUtils.singleSignOutDisabled
@@ -83,8 +87,8 @@ security.shiro.cas.singleSignOut.logoutParameterName = "slo"
         and: "the configured values are used"
         ShiroCasConfigUtils.serverUrl == "https://cas.example.com"
         ShiroCasConfigUtils.serviceUrl == "https://app.example.com/shiro-cas"
-        ShiroCasConfigUtils.loginUrl == "https://cas.example.com/customLogin?renew=true"
-        ShiroCasConfigUtils.logoutUrl == "https://cas.example.com/customLogout"
+        ShiroCasConfigUtils.defaultLoginUrl == "https://cas.example.com/customLogin?renew=true"
+        ShiroCasConfigUtils.defaultLogoutUrl == "https://cas.example.com/customLogout"
         ShiroCasConfigUtils.failureUrl == "https://app.example.com/casFailure"
         ShiroCasConfigUtils.shiroCasFilter == "/shiro-cas=singleSignOutFilter,casFilter\n/other=otherFilter"
         !ShiroCasConfigUtils.singleSignOutDisabled
@@ -102,8 +106,8 @@ security.shiro.cas.serviceUrl = "http://localhost:8080/app/shiro-cas/"
         then: "the trailing slash is ignored"
         ShiroCasConfigUtils.serverUrl == "https://localhost/cas"
         ShiroCasConfigUtils.serviceUrl == "http://localhost:8080/app/shiro-cas"
-        ShiroCasConfigUtils.loginUrl == "https://localhost/cas/login?service=http://localhost:8080/app/shiro-cas"
-        ShiroCasConfigUtils.logoutUrl == "https://localhost/cas/logout?service=http://localhost:8080/app/shiro-cas"
+        ShiroCasConfigUtils.defaultLoginUrl == "https://localhost/cas/login?service=http://localhost:8080/app/shiro-cas"
+        ShiroCasConfigUtils.defaultLogoutUrl == "https://localhost/cas/logout?service=http://localhost:8080/app/shiro-cas"
     }
 
     void "specified login parameters are honored"() {
@@ -124,9 +128,9 @@ security.shiro.cas.loginParameters.welcome = "Welcome to Shiro Cas"
         ShiroCasConfigUtils.serviceUrl == "http://localhost:8080/app/shiro-cas"
 
         and: "other values are either defaulted or based on the configured values"
-        ShiroCasConfigUtils.loginUrl ==
+        ShiroCasConfigUtils.defaultLoginUrl ==
                 "https://localhost/cas/login?service=http://localhost:8080/app/shiro-cas&renew=true&gateway=true&welcome=Welcome%20to%20Shiro%20Cas"
-        ShiroCasConfigUtils.logoutUrl == "https://localhost/cas/logout?service=http://localhost:8080/app/shiro-cas"
+        ShiroCasConfigUtils.defaultLogoutUrl == "https://localhost/cas/logout?service=http://localhost:8080/app/shiro-cas"
         ShiroCasConfigUtils.failureUrl == null
         ShiroCasConfigUtils.shiroCasFilter == "/shiro-cas=singleSignOutFilter,casFilter\n"
     }
@@ -145,6 +149,41 @@ security.shiro.cas.singleSignOut.disabled = true
         and: "single sign out support is disabled"
         ShiroCasConfigUtils.singleSignOutDisabled
         ShiroCasConfigUtils.shiroCasFilter == "/shiro-cas=casFilter\n"
+    }
+
+    void "enabling isServerNameDynamic handles multiple server names"(){
+        setup:
+        def firstUrl = "http://test.server.com"
+        def secondUrl = "http://anothertest.server.com"
+
+        GroovyMock(SecurityUtils, global: true)
+        GroovyMock(WebUtils, global: true)
+        def httpServletRequest = Mock(HttpServletRequest)
+        WebUtils.getHttpRequest(_) >> httpServletRequest
+
+
+        when: "initialized with a dynamicServerName configuration"
+        init("""
+security.shiro.cas.serverUrl = "https://localhost/cas"
+security.shiro.cas.serviceUrl = "http://localhost:8080/app/shiro-cas"
+security.shiro.cas.servicePath = "/test/shiro-cas"
+security.shiro.cas.failurePath = "/test/"
+        """)
+
+        def firstServiceUrl = ShiroCasConfigUtils.serviceUrl
+        def firstFailureUrl = ShiroCasConfigUtils.failureUrl
+        def secondServiceUrl = ShiroCasConfigUtils.serviceUrl
+        def secondFailureUrl = ShiroCasConfigUtils.failureUrl
+
+        then: "URLs overridden using first domain"
+        2 * httpServletRequest.getRequestURL() >> new StringBuffer(firstUrl)
+        firstServiceUrl == firstUrl + "/test/shiro-cas"
+        firstFailureUrl == firstUrl + "/test/"
+
+        then: "URLs overridden using second domain"
+        2 * httpServletRequest.getRequestURL() >> new StringBuffer(secondUrl)
+        secondServiceUrl == secondUrl + "/test/shiro-cas"
+        secondFailureUrl == secondUrl + "/test/"
     }
 
     static void init(String script) {